+++ packages/radix-ui-themes/src/components/tabs.props.ts
@@ -1,11 +1,15 @@
-import { PropDef } from '../helpers';
+import { PropDef, colorProp, highContrastProp } from '../helpers';
 
 const sizes = ['1', '2'] as const;
 
 const tabsListPropDefs = {
   size: { type: 'enum', values: sizes, default: '2', responsive: true },
+  color: { ...colorProp, default: undefined },
+  highContrast: highContrastProp,
 } satisfies {
   size: PropDef<(typeof sizes)[number]>;
+  color: typeof colorProp;
+  highContrast: typeof highContrastProp;
 };
 
 export { tabsListPropDefs };