+++ packages/radix-ui-themes/changelog.md
@@ -9,6 +9,8 @@
   - Change the size of the bounding box to match the size of the Slider track
 - `Skeleton`: New component
 - `Spinner`: New component
+- `Tabs`:
+  - Add `color` and `highContrast` props to `TabsList`
 
 ## 2.0.3